+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 #### Updated
 - Updating to React 16 and Enzyme 3.9 with updates to unit tests.
 
+### v1.1.6
+#### Added
+- Created a new tab to display information about each library's focus area and service area.
+
+### v1.1.5
+#### Added
+- Implemented a toggle for switching between production mode and QA mode.
+
 ### v1.1.4
 #### Updated
 - Utilized new built-in styling options for the Button component.